+++ .storybook/preview.js
@@ -4,7 +4,6 @@ export const decorators = [
   (Story) => (
     <div
       style={{
-        background: 'white',
         padding: '20px',
         boxSizing: 'border-box',
         height: '400px',