+++ src/components/Sparkbar/stories/Sparkbar.stories.tsx
@@ -13,7 +13,7 @@ export default {
   component: Sparkbar,
   decorators: [
     (Story: any) => (
-      <div style={{width: '200px', height: '200px'}}>{Story()}</div>
+      <div style={{width: '200px', height: '100px'}}>{Story()}</div>
     ),
   ],
   argTypes: {
@@ -49,5 +49,27 @@ Default.args = {
     {x: 10, y: comparisonValue},
   ],
   color: primaryColor,
+  barFillStyle: 'solid',
+};
+
+export const OffsetAndNulls = Template.bind({});
+OffsetAndNulls.args = {
+  data: [100, 200, -300, null, 400, 0, 0, 400, 700, 900, 500],
+  dataOffsetRight: 12,
+  dataOffsetLeft: 50,
+  comparison: [
+    {x: 0, y: comparisonValue},
+    {x: 1, y: comparisonValue},
+    {x: 2, y: comparisonValue},
+    {x: 3, y: comparisonValue},
+    {x: 4, y: comparisonValue},
+    {x: 5, y: comparisonValue},
+    {x: 6, y: comparisonValue},
+    {x: 7, y: comparisonValue},
+    {x: 8, y: comparisonValue},
+    {x: 9, y: comparisonValue},
+    {x: 10, y: comparisonValue},
+  ],
+  color: 'darkModePositive',
   barFillStyle: 'solid',
 };