+++ src/types.ts
@@ -131,3 +131,5 @@ export interface Dimensions {
   width: number;
   height: number;
 }
+
+export type SparkChartData = number | null;